Changes to allow clang to compile the gnu tools.
git-svn-id: http://ellcc.org/svn/ellcc/trunk@1688 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ gnu/src/bfd/coffcode.h
@@ -1065,7 +1065,7 @@ styp_to_sec_flags (bfd *abfd,
 	case IMAGE_SCN_MEM_EXECUTE:
 	  sec_flags |= SEC_CODE;
 	  break;
-	case IMAGE_SCN_MEM_WRITE:
+	case (long)IMAGE_SCN_MEM_WRITE:
 	  sec_flags &= ~ SEC_READONLY;
 	  break;
 	case IMAGE_SCN_MEM_DISCARDABLE:
